+++ CHANGELOG.md
@@ -10,6 +10,11 @@
     * `exportKey`
     * `getAccount`
     * `listAccounts`
+    * `createSubnet`
+    * `platform.getSubnets`
+    * `platform.validatedBy`
+    * `platform.validates`
+    * `platform.getBlockchains`
 * Added tests for the following non-PChain endpoints:
     * `admin.getNodeID`
     * `admin.peers`